+++ commands/themelio-client/src/common/formatter/formatter.rs
