+++ commands/themelio-client/src/common/formatter/mod.rs
@@ -1,2 +1,3 @@
 pub mod json;
 pub mod plain;
+pub mod formatter;
