+++ commands/themelio-client/src/common/formatter/plain.rs
@@ -8,7 +8,7 @@ use blkstructs::{CoinDataHeight, CoinID};
 
 use crate::wallet::data::WalletData;
 use crate::wallet::wallet::Wallet;
-use crate::common::formatter::OutputFormatter;
+use crate::common::formatter::formatter::OutputFormatter;
 
 pub struct PlainOutputFormatter {
