+++ commands/themelio-client/src/common/mod.rs
@@ -2,5 +2,4 @@ pub mod context;
 pub mod executor;
 pub mod input;
 pub mod output2;
-pub mod output;
 pub mod formatter;
