+++ commands/themelio-client/src/wallet_utils/executor.rs
@@ -3,7 +3,7 @@ use blkstructs::{CoinDataHeight, Transaction};
 use crate::common::context::ExecutionContext;
 use crate::common::executor::CommonCommandExecutor;
 use crate::wallet::manager::WalletManager;
-use crate::common::formatter::OutputFormatter;
+use crate::common::formatter::formatter::OutputFormatter;
 
 /// Responsible for executing a single client CLI command non-interactively.
 pub struct WalletUtilsExecutor {
